+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/aries/PresentProofMessenger.kt
@@ -1,5 +1,7 @@
-package at.asitplus.wallet.lib.agent
+package at.asitplus.wallet.lib.aries
 
+import at.asitplus.wallet.lib.agent.Holder
+import at.asitplus.wallet.lib.agent.Verifier
 import at.asitplus.wallet.lib.data.ConstantIndex
 import com.benasher44.uuid.uuid4
